agents/datagema/requirements.txt
torch
data-gemma
transformers
gradio

torch
data-gemma
google-colab
transformers
textwrap
ipython
gradio